+++ history.md
@@ -3,6 +3,7 @@
 - Remove `disableWorkspaceFolders` `disableDiagnostics` and `disableCompletion`
   from language client option.
 - Add configuration `documentHighlight.timeout`.
+- Add `tabPersist` option to `ListAction`.
 
 # 2022-01-30